Scroll wheel to change fog depth doesn't move cam

+++ src/ui/md.rs
@@ -278,7 +278,7 @@ pub fn md_setup(
 
                 ui.label("E / atom: ");
                 // todo: Don't continuosly run this!
-                let e_per_atom = (snap.energy_kinetic / ((snap.water_o_posits.len() * 3) as f32 + snap.atom_posits.len() as f32)) as u16;
+                let e_per_atom = snap.energy_kinetic / ((snap.water_o_posits.len() * 3) as f32 + snap.atom_posits.len() as f32);
                 ui.label(RichText::new(format!("{:.1}", e_per_atom)).color(Color32::GOLD));
 
                 ui.label("PE: ");
